Improved code blocks in examples, centralised theme files, and improved other CSS styles.
